+++ RentACar.API/Controllers/CarImagesController.cs
@@ -23,9 +23,9 @@ namespace RentACar.WebAPI.Controllers
         {
             var result = _carImageService.GetAllCarImages();
 
-            if (!result.Success)
+            if (!result.Success || result.Data == null || !result.Data.Any())
             {
-                return BadRequest(result);
+                return NotFound(result.Message);
             }
 
             return Ok(result);
@@ -34,6 +34,11 @@ namespace RentACar.WebAPI.Controllers
         [HttpGet("getbyid")]
         public async Task<IActionResult> GetByIdAsync(int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest("The provided ID must be a positive number.");
+            }
+
             var result = await _carImageService.GetSingleAsync(id);
 
             if (!result.Success)
@@ -77,13 +82,23 @@ namespace RentACar.WebAPI.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteAsync([FromQuery] int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest("The provided ID must be a positive number.");
+            }
+
             var carImageToDelete = await _carImageService.GetSingleAsync(id);
+
+            if (!carImageToDelete.Success)
+            {
+                return BadRequest(carImageToDelete);
+            }
 
             var result = await _carImageService.DeleteAsync(carImageToDelete.Data);
 
             if (!result.Success)
             {
-                return BadRequest(result);
+                return NotFound(result);
             }
 
             return Ok(result);
