+++ RentACar.API/Controllers/CarsController.cs
@@ -21,9 +21,9 @@ namespace RentACar.WebAPI.Controllers
         {
             var result = _carService.GetAllCars();
 
-            if (!result.Success)
+            if (!result.Success || result.Data == null || !result.Data.Any())
             {
-                return BadRequest(result);
+                return NotFound(result.Success);
             }
 
             return Ok(result);
@@ -32,11 +32,16 @@ namespace RentACar.WebAPI.Controllers
         [HttpGet("getbyid")]
         public async Task<IActionResult> GetByIdAsync(int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest("The provided ID must be a positive number.");
+            }
+
             var result = await _carService.GetSingleAsync(id);
 
             if (!result.Success)
             {
-                return BadRequest(result.Success);
+                return NotFound(result.Success);
             }
 
             return Ok(result);
@@ -50,7 +55,7 @@ namespace RentACar.WebAPI.Controllers
 
             if (!result.Success)
             {
-                return BadRequest(result);
+                return NotFound(result.Message);
             }
 
             return Ok(result);
@@ -75,13 +80,24 @@ namespace RentACar.WebAPI.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteAsync([FromQuery] int id)
         {
+
+            if (id <= 0)
+            {
+                return BadRequest("The provided ID must be a positive number.");
+            }
+
             var carToDelete = await _carService.GetSingleAsync(id);
+
+            if (!carToDelete.Success)
+            {
+                return NotFound(carToDelete);
+            }
 
             var result = await _carService.DeleteAsync(carToDelete.Data);
 
             if (!result.Success)
             {
-                return BadRequest(result);
+                return NotFound(result);
             }
 
             return Ok(result);
