+++ RentACar.API/Controllers/InvoicesController.cs
@@ -22,9 +22,10 @@ namespace RentACar.WebAPI.Controllers
         {
             var result = _invoiceService.GetAllInvoices();
 
-            if (!result.Success)
+            if (!result.Success || result.Data == null || !result.Data.Any())
             {
-                return BadRequest(result);
+                return NotFound(result.Message);
+
             }
 
             return Ok(result);
@@ -33,11 +34,16 @@ namespace RentACar.WebAPI.Controllers
         [HttpGet("getbyid")]
         public async Task<IActionResult> GetById(int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest("The provided ID must be a positive number.");
+            }
+
             var result = await _invoiceService.GetSingleAsync(id);
 
             if (!result.Success)
             {
-                return BadRequest(result.Success);
+                return NotFound(result.Success);
             }
 
             return Ok(result);
@@ -76,13 +82,24 @@ namespace RentACar.WebAPI.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteAsync([FromQuery] int id)
         {
+
+            if (id <= 0)
+            {
+                return BadRequest("The provided ID must be a positive number.");
+            }
+
             var invoiceToDelete = await _invoiceService.GetSingleAsync(id);
+
+            if (!invoiceToDelete.Success)
+            {
+                return NotFound(invoiceToDelete.Success);
+            }
 
             var result = await _invoiceService.DeleteAsync(invoiceToDelete.Data);
 
             if (!result.Success)
             {
-                return BadRequest(result);
+                return NotFound(result);
             }
 
             return Ok(result);
