+++ RentACar.API/Controllers/LocationsController.cs
@@ -23,9 +23,10 @@ namespace RentACar.WebAPI.Controllers
 
             var result = _locationService.GetAllLocations();
 
-            if (!result.Success)
+            if (!result.Success || result.Data == null || !result.Data.Any())
             {
-                return BadRequest(result.Message);
+                return NotFound(result.Message);
+
             }
 
             return Ok(result);
@@ -35,11 +36,16 @@ namespace RentACar.WebAPI.Controllers
         [HttpGet("getbyid")]
         public async Task<IActionResult> GetById(int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest("The provided ID must be a positive number.");
+            }
+
             var result = await _locationService.GetSingleAsync(id);
 
             if (!result.Success)
             {
-                return BadRequest(result);
+                return NotFound(result);
             }
 
             return Ok(result);
@@ -79,13 +85,23 @@ namespace RentACar.WebAPI.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteAsync([FromQuery] int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest("The provided ID must be a positive number.");
+            }
+
             var locationToDelete = await _locationService.GetSingleAsync(id);
+
+            if (!locationToDelete.Success)
+            {
+                return NotFound(locationToDelete);
+            }
 
             var result = await _locationService.DeleteAsync(locationToDelete.Data);
 
             if (!result.Success)
             {
-                return BadRequest(result);
+                return NotFound(result);
             }
 
             return Ok(result);
