+++ RentACar.API/Controllers/TransmissionsController.cs
@@ -21,9 +21,10 @@ namespace RentACar.WebAPI.Controllers
         {
             var result = _transmissionService.GetAllTransmissions();
 
-            if (!result.Success)
+            if (!result.Success || result.Data == null || !result.Data.Any())
             {
-                return BadRequest(result);
+                return NotFound(result.Message);
+
             }
 
             return Ok(result);
@@ -32,11 +33,16 @@ namespace RentACar.WebAPI.Controllers
         [HttpGet("getbyid")]
         public async Task<IActionResult> GetByIdAsync(int id)
         {
+            if(id <= 0)
+            {
+                return BadRequest("The provided ID must be a positive number.");
+            }
+
             var result = await _transmissionService.GetSingleAsync(id);
 
             if (!result.Success)
             {
-                return BadRequest(result.Success);
+                return NotFound(result.Success);
             }
 
             return Ok(result);
@@ -75,13 +81,23 @@ namespace RentACar.WebAPI.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteAsync([FromQuery] int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest("The provided ID must be a positive number.");
+            }
+
             var transmissionToDelete = await _transmissionService.GetSingleAsync(id);
+
+            if (!transmissionToDelete.Success)
+            {
+                return NotFound(transmissionToDelete);
+            }
 
             var result = await _transmissionService.DeleteAsync(transmissionToDelete.Data);
 
             if (!result.Success)
             {
-                return BadRequest(result);
+                return NotFound(result);
             }
 
             return Ok(result);
